+++ src/Lab1/vehicles/Scania.java
@@ -1,4 +1,4 @@
-package Lab1;
+package Lab1.vehicles;
 
 import java.awt.*;
