+++ src/Lab1/vehicles/Vehicle.java
@@ -1,7 +1,8 @@
-package Lab1;
+package Lab1.vehicles;
 
 import java.awt.*;
-import Lab1.interfaces.IMovable;
+
+import Lab1.vehicles.interfaces.IMovable;
 
 public abstract class Vehicle implements IMovable {
     protected int nrDoors; // Number of doors on the car
