+++ src/Lab1/vehicles/Volvo240.java
@@ -1,4 +1,4 @@
-package Lab1;
+package Lab1.vehicles;
 
 import java.awt.*;
